chat-stream/python-backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
asyncpg>=0.30.0
python-dotenv>=1.0.0
httpx==0.27.0
sse-starlette==1.6.5
black>=22.0.0
